Convert to use new MPCoreSystem
