+++ lua/ge/extensions/UI.lua
@@ -116,7 +116,7 @@ end
 
 local function chatSend(msg)
 	local c = 'C:'..MPConfig.getNickname()..": "..msg
-	MPGameNetwork.send(c)
+	MPCoreSystem.send('GAME',c)
 	TriggerClientEvent("ChatMessageSent", c)
 end
