+++ lua/vehicle/extensions/BeamMP/MPVehicleVE.lua
@@ -48,7 +48,7 @@ local function onKeyStateChanged(key, state)
 			keypressTriggers[i].func(state)
 		end
 	end
-	obj:queueGameEngineLua("MPGameNetwork.onKeyStateChanged('"..key.."',"..tostring(state)..")")
+	obj:queueGameEngineLua("MPCoreSystem.onKeyStateChanged('"..key.."',"..tostring(state)..")")
 end
 
 function getKeyState(key)
